+++ src/components/Container/PriceContainer/PriceContainer.style.tsx
@@ -2,7 +2,6 @@ import styled from 'styled-components';
 
 export const PriceContainer = styled.div`
   display: flex;
-  flex-direction: row;
   align-items: center;
   justify-content: space-between;
   padding: 5px 0;
